NewFeature：新增 Flink 实现分组 TopN

+++ module-flink/src/main/java/com/dream/flink/data/Order.java
@@ -22,7 +22,7 @@ public class Order {
     int goodsId;
 
     /** 价格 */
-    int price;
+    long price;
 
     /** 城市 */
     int cityId;
